Move image_selection and the stitcher under agent/tools
Every consumer of image_selection was already in agent/: surface_views backs
render (via wall_refs) and select_views (via overlay), and surface_compare/
backs the render source. select_references.py and object_views.py are reached
by nothing at all. room_format was holding the agent's tooling.

The shared backing moves with it. config, scan and overlay were sitting at the
agent/ top level, but nothing outside the tools uses them — only the tools and
the pipeline's evidence.py — so they belong under agent/tools/shared/ with
image_selection and the stitcher.

This frees select_views.py, which had to stay behind last round only because
select_references.py spawns it by path from what was then room_format. Both are
under agent/tools now, so it finally sits beside select_for.py and the
select_views tool owns its whole source.

Four sys.path inserts become real imports: overlay and wall_refs each reached
stitch_wall and surface_views that way, and surface_views reached the stitcher.
They were path hacks only because the code lived somewhere the tools could not
import from.

render_ortho.py needed the same fix as _blender_frames.py last commit — it
joined "../../room_render", a sibling relationship this move broke. It anchors
on the package directory by name now.

What remains in room_format/rendering is room-format work: room_render/ and the
object turntables.
