+++ src/litereality_agent/agent/tools/select_views/source/select_for.py
@@ -23,10 +23,10 @@ import sys
 
 from litereality_agent import REPO_ROOT
 
-# `select_views` stays in room_format: `image_selection/select_references.py` spawns it by path to
-# build the reference sheets the init stage consumes, and moving it here would point room_format at
-# agent. A real package import replaces the `sys.path` insert this used when the two were siblings.
-from litereality_agent.room_format.rendering.room_render import select_views as SV
+# `select_views` now sits beside this file. `image_selection/select_references.py` still spawns it by
+# path to build the reference sheets init consumes; both live under `agent/tools` now, so that is an
+# intra-package reference. A real import replaces the `sys.path` insert this used to need.
+from litereality_agent.agent.tools.select_views.source import select_views as SV
 
 
 def views_for(manifest: dict, target_id: str, n: int = 3, min_quality: float = 0.5) -> list[dict]:
