+++ src/litereality_agent/agent/tools/shared/image_selection/README.md
