+++ src/litereality_agent/agent/tools/shared/image_selection/surface_compare/README.md
