+++ src/litereality_agent/agent/tools/shared/stitch_wall_image/stitch_wall.py
